+++ docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 **If you need help updating/installing or have found any bugs, please join [our Discord server](https://discord.gg/eGnrPqEH7U) or open a [GitHub issue here](https://github.com/Dev-CasperTheGhost/snaily-cadv3/issues/new/choose)**
 
+## 2.0.5
+
+- Minor: fixed 404 error with Call Events
+- Minor: fixed minor bug in values management
+
 ## 2.0.4
 
 - New: able to remove citizen image